Add more explicit casts and remove dead code
Comment unused SERVER from DeviceShutdown.
Comment unused max from parse_hostport.
Comment unused nodeptr from ixmlNode_cloneDoc.
Comment unused newNode from Parser_hasDefaultNamespace.
Comment unused Parser_parseReference function
Check return code of shutdown and display an error if needed.

(forward port of commit cb07623ddece5c5cd1f2867336d92aee938a1334)

+++ upnp/src/genlib/net/http/httpparser.c
@@ -251,7 +251,7 @@ static parse_status_t scanner_get_token(
 			/* possibly more valid chars */
 			return PARSE_INCOMPLETE;
 		/* calc token length */
-		token->length = (size_t)(cursor - token->buf);
+		token->length = (size_t)cursor - (size_t)token->buf;
 	} else if (c == ' ' || c == '\t') {
 		token->buf = cursor++;
 		token_type = TT_WHITESPACE;
@@ -260,7 +260,7 @@ static parse_status_t scanner_get_token(
 		if (!scanner->entire_msg_loaded && cursor == null_terminator)
 			/* possibly more chars */
 			return PARSE_INCOMPLETE;
-		token->length = (size_t)(cursor - token->buf);
+		token->length = (size_t)cursor - (size_t)token->buf;
 	} else if (c == TOKCHAR_CR) {
 		/* scan CRLF */
 		token->buf = cursor++;
@@ -306,7 +306,7 @@ static parse_status_t scanner_get_token(
 				return PARSE_FAILURE;
 		}
 		if (got_end_quote)
-			token->length = (size_t)(cursor - token->buf);
+			token->length = (size_t)cursor - (size_t)token->buf;
 		else {	/* incomplete */
 
 			assert(cursor == null_terminator);
@@ -1416,7 +1416,7 @@ parse_status_t parser_parse_responseline(INOUT http_parser_t *parser)
 	while (*p == ' ' || *p == '\t')
 		p++;
 	/* now, p is at start of status msg */
-	n = line.length - (size_t)(p - line.buf);
+	n = line.length - ((size_t)p - (size_t)line.buf);
 	if (membuffer_assign(&hmsg->status_msg, p, n) != 0) {
 		/* out of mem */
 		parser->http_error_code = HTTP_INTERNAL_SERVER_ERROR;
@@ -1737,7 +1737,7 @@ parser_parse_chunky_entity( INOUT http_parser_t * parser )
                       ( scanner->cursor - save_pos ) );
     scanner->cursor = save_pos; /* adjust scanner too */
 
-    if( parser->chunk_size == 0 ) {
+    if( parser->chunk_size == (size_t)0 ) {
         /* done reading entity; determine length of entity */
         parser->msg.entity.length = parser->scanner.cursor -
             parser->entity_start_position + parser->msg.amount_discarded;
